refactor; move functions to util module
